--- cvORBNetStream.cpp
@@ -1,53 +1,0 @@
-#include "cvORBNetStream.h"
-#include <iostream>
-
-cvORBNetStream::cvORBNetStream()
-{
-  // Set up the ZMQ context
-  context = zmq::context_t(1);
-}
-
-cvORBNetStream::~cvORBNetStream()
-{
-  socket.close();
-}
-
-/**
- * Quick note regarding both Init functions: You might notice that Init() can be called however many times
- * you want, this is because I've yet to thought of a way to make a proper C++ class that has a nice
- * factory method or perhaps different subclasses; am not used with OOP with C++ yet and I don't intend
- * to spend too much time on this.
- */
-
-int cvORBNetStream::Init(int port)
-{
-  std::cout << "Initializing worker stream on port " << port << std::endl;
-  assert(this->port == 0); // Make sure the port hasn't been set yet (can't reinitialize)
-  this->port = port;
-  socket = zmq::socket_t(context, zmq::socket_type::rep);
-  socket.bind("tcp://*:" + std::to_string(port));
-  return 0;
-}
-
-int cvORBNetStream::SendFrame(std::string encodedFrame)
-{
-  zmq::message_t request(encodedFrame.size());
-  memcpy(request.data(), encodedFrame.c_str(), encodedFrame.size());
-  std::cout << "sending frame .." << std::endl;
-
-  // TODO: What if there's no request? Add a buffer with a safe-distancing semaphore.
-  try
-  {
-    zmq::message_t temp;
-    socket.recv(temp, zmq::recv_flags::none);
-    socket.send(request, zmq::send_flags::none);
-  }
-  catch (zmq::error_t e)
-  {
-    std::cout << "Error sending frame: " << e.what() << std::endl;
-    return -1;
-  }
-
-  return 0;
-}
-
